MCOL-1078 Fix packet stitching
We allocate 64KB memory before doing a read and assume the read will fill
the 64KB before the next read when the packet is > 64KB. This is not
always the case, the kernel network buffer may not have 64KB in it yet.
This caused holes when the network packet was stitched back together. We
now keep track of how much of the buffer we have filled.