MCOL-1245 Fix decompression double-free
If the decompression buffer was used twice a double-free event would
occur causing a segfault. This fixes that problem and an error in the
compiling documentation.

No test since it requires a remote connection to trigger. You need a
table with a large amount of columns so that the decompression will be
used for the read of the system catalog and the HWM data.