MCOL-1816 Add char* setters
The compiler was sending char* setters to bool instead of std::string as
it is easier to evaluate a pointer to a boolean.

I've started adding boost::string_ref to pass arbitrary strings around
which now allows us to use char* setters. It also may improve memory
usage / performance slightly.

We should look at replacing our string usage with this in more places in
the future.